pyproject.toml
[tool.poetry]
name = "city-bike-app"
version = "0.0.1"
description = ""
authors = ["Ilkka Kivistö <ilkka.kivisto@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^2.0.0"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
psycopg2 = "^2.9.5"


[tool.poetry.group.dev.dependencies]
coverage = "^7.0.5"
pytest = "^7.2.1"
pylint = "^2.15.10"
autopep8 = "^2.0.1"
robotframework = "^6.0.2"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
